Move CLI config initialization
Wrap CLI configuration in a class that returns the configured values.
