+++ index.js
@@ -2,72 +2,23 @@ import fs from "fs";
 import path from "path";
 import { BrowserFactory, CONNECTION, factoryOptions } from "./src/TBBrowserFactory";
 import {CapabilityFactory} from "./src/TBCapabilityFactory";
-import testFunctions from "./src/test_functions";
 import {ConfigurationParser} from "./src/ConfigurationParser";
 import {createImageWriter} from "./src/writeImageData";
-import meow from 'meow';
-
 import {TBBatchRunner} from "./src/TBRunTestInBatches";
+import {CliRequestFactory} from "./src/CliRrequestFactory";
 
 const METADATA_FILENAME = 'metadata.json';
 
-const cli = meow(
-	`
-	Usage
-		node -r esm index.js [OPTION...] <CAMPAIGN_NAME>
+const requestFactory = new CliRequestFactory( __dirname );
 
-		<CAMPAIGN_NAME>			Name of the campaign to create screenshots of
-		
-	Options
-		--screenshotPath, -s 		Path to directory containing campaign directories with metadata, default: "banner-shots"
-		--configName, -c		Path to the campaign config file (toml), default: "campaign_info.toml"
-		--concurrentRequestLimit, -l 	Amount of concurrent requests to saucelab, default: 4
-		--testFunctionName, -t		Name of the test function, default: "shootBanner"
-		
-	Examples
-	$ node -r esm index.js desktop
-	`,
-	{
-		description: 'Create screenshots of a certain banner campaign',
-		input: [],
-		flags: {
-			screenshotPath: {
-				alias: 's',
-				type: 'string',
-				default: 'banner-shots'
-			},
-			configName: {
-				alias: 'c',
-				type: 'string',
-				default: 'campaign_info.toml'
-			},
-			concurrentRequestLimit: {
-				alias: 'l',
-				type: 'number',
-				default: 5
-			},
-			testFunctionName: {
-				alias: 't',
-				type: 'string',
-				default: 'shootBanner'
-			}
-		}
-	} );
-if (typeof cli.input[0] !== 'string' ) {
-	console.log( `ERROR: Please provide an existing campaign name! \nSee --help for usage instructions.` );
-	process.exit( 2 );
-}
-const campaignName = cli.input[0];
-const screenshotPath = path.join( __dirname, cli.flags.screenshotPath.toString() );
-const configName = path.join( __dirname, cli.flags.configName.toString() );
-const concurrentRequestLimit = cli.flags.concurrentRequestLimit;
-const testFunctionName = cli.flags.testFunctionName;
+const {
+	campaignName,
+	configName,
+	screenshotPath,
+	concurrentRequestLimit,
+	testFunction
+} = requestFactory.getRequestParameters();
 
-if (typeof testFunctions[testFunctionName] !== 'function' ) {
-	console.log( `Unknown test function "${ testFunctionName }"` );
-	process.exit( 2 );
-}
-const shootBanner = testFunctions[testFunctionName];
 const config = fs.readFileSync( configName );
 const parser = new ConfigurationParser( config.toString() );
 const testCaseGenerator = parser.generate( campaignName );
@@ -87,15 +38,15 @@ const outputDirectory = path.join( screenshotPath, parser.getCampaignTracking( c
 	 * @param {Browser} browser
 	 * @returns {Promise}
 	 */
-	const shoot = async (testCase, browser ) => {
+	const testFunctionWithImageWriter = async (testCase, browser ) => {
 		try {
-			await shootBanner( browser, testCase, imageWriter );
+			await testFunction( browser, testCase, imageWriter );
 		} catch( e ) {
 			console.log( "browser error", e );
 		}
 	};
 
-	await batchRunner.runTestsInBatches(concurrentRequestLimit, testCases, shoot);
+	await batchRunner.runTestsInBatches(concurrentRequestLimit, testCases, testFunctionWithImageWriter);
 
 	testCases.map( testcase => {
 		console.log( testcase.state.description, testcase.getName() );
